requirements.txt
fastapi
uvicorn[standard]
pydantic-ai
sqlmodel
polyfactory
python-dotenv
ipykernel 
pandas
matplotlib
aiofiles